+++ src/test/static/polymer2/test-element-17.js
@@ -1,0 +1,18 @@
+/**
+ * @customElement
+ * @polymer
+ * @extends Polymer.Element
+ */
+class BaseElement extends Polymer.Element {
+  static get properties() {
+    return {
+      foo: {
+        notify: true,
+        type: String,
+        reflectToAttribute: true,
+        readOnly: true,
+        value: 'foo'
+      },
+    };
+  }
+}